requirements.txt
accelerate
easydict
einops
fvcore
hydra-core
imageio
nltk
numpy
omegaconf
openai-clip
opencv-python
pandas
peft==0.12.0
plyfile
scipy
sentence-transformers
sentencepiece
submitit
timm
tokenizers
transformers==4.42.4
transforms3d
wandb
